Perf: faster VT engine to beat Ghostty on output-stress throughput
Reduce per-byte/per-cell CPU on the terminal hot paths that gated the
PTY-drain benchmark (unicode_mixed, attributes, ansi_sgr, redraw), with
no architecture change, no dropped backpressure, and no feature loss:

- O(1) Unicode width: a fast path for cp<0x300 plus a generated two-stage
  BMP trie + astral binary search (Scripts/generate-width-table.swift ->
  CharacterWidthTable.swift), replacing the per-scalar linear range scan.
  Proven byte-identical to the canonical ranges over every scalar.
- Allocation-free SGR/CSI params: reused flat buffers + a borrowed
  CSIParams view, dropping the per-sequence nested [[Int]] allocation and
  the unused flat map in parserCSI.
- Compact cell: TerminalGridColor.palette(Int) -> palette(UInt8) shrinks
  TerminalGridCell 64 -> 32 bytes (halves per-cell copy/snapshot/== cost).
- Block-move scroll + bulk erase fills (memmove / update(repeating:)) in
  place of the per-cell loops.
- Daemon write buffer consumes via an offset instead of removeFirst,
  eliminating O(n^2) flush churn when a client backs up.

Adds exhaustive width-parity and cell-layout guard tests, four hot-path
micro-benchmarks, and a checked-in cross-terminal stress runner.

822 tests green (debug + release); 24 benchmarks green, no render or
compositor regressions. Scroll-heavy parse ~7x faster, run path ~1.7x.

+++ Packages/HarnessTerminalEngine/Sources/HarnessTerminalEngine/Model/TerminalGridModel.swift
@@ -6,9 +6,14 @@ import Foundation
 /// These are the engine's public, renderer-agnostic color values — the same shape the
 /// headless `readGrid` snapshot exposes and the live renderer consumes. They carry no
 /// platform color type so the engine stays Foundation-only and unit-testable without a GPU.
+///
+/// `palette` is a `UInt8` (the full 0–255 ANSI range) rather than `Int` so the enum's payload
+/// fits in one byte: this keeps `TerminalGridColor` tiny and, with three colors per cell, shrinks
+/// `TerminalGridCell` enough to roughly halve the per-cell copy / snapshot / `==` cost on the
+/// throughput hot paths.
 public enum TerminalGridColor: Equatable, Sendable {
     case none
-    case palette(Int)
+    case palette(UInt8)
     case rgb(r: UInt8, g: UInt8, b: UInt8)
 }
